+++ example/lib/main.dart
@@ -2,5 +2,5 @@ import 'package:flutter/material.dart';
 import 'package:well_formed/well_formed.dart';
 
 void main() {
-  runApp(FormFieldTester(BasicTextField(blank: 'Please fill in this field')));
+  runApp(FormFieldTester(BasicTextField(trim: true, blank: 'Please fill in this field')));
 }