+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ## [Unreleased]
 
-## [0.0.1] - 2021-07-21
-
 ### Added
 
 - BasicTextField, a text form field that can be made required and/or have its
-  input data trimmed — [8](https://github.com/dartoos-deve/well_formed/issues)
+  input data trimmed — [12](https://github.com/dartoos-dev/well_formed/issues/12).
+- DigitField, a digit-only form field —
+  [11](https://github.com/dartoos-dev/well_formed/issues/11).
+- Example code that can be run —
+  [9](https://github.com/dartoos-dev/well_formed/issues/9).
+
+## [0.0.1] - 2021-07-21